Prep for work fixing all the buttons
