+++ csfieldguide/templates/interactives/box-translation.html
@@ -4,8 +4,8 @@
 {% load static %}
 
 {% block html %}
-  <div class="row mb-3 box-translation">
-    <div id="container">
+  <div class="row mb-3">
+    <div id="container" class="box-translation">
       <button onClick="reset()" id="restart-button">Restart</button>
       <div id="user-input" class="desktop-layout">
         <div id="coordinates">
